requirements.txt
fastapi
uvicorn
httpx
pytest
pillow
imageio
